refactor: rename components folder to features
