--- task-1/src/components/index.ts
@@ -1,7 +1,0 @@
-export * from './error-boundries';
-export * from './error-button-wrapper';
-export * from './pagination';
-export * from './loader';
-export * from './contexts';
-export * from './theme-button-wrapper';
-export * from './drawer';
