+++ task-1/src/features/items-selector/index.ts
@@ -1,1 +1,1 @@
-export * from './itemsSelector';
+export * from './items-selector';
