+++ task-1/src/features/items-selector/items-selector.tsx
