+++ task-1/src/features/loader/loader.tsx
