+++ task-1/src/features/pagination/style.css
