+++ task-1/src/hooks/use-charactes-close.ts
